Replaced zoxide+tmux with primeagen's tmux-sessionizer
Using zoxide and the tmux plugin to create sessions with it was pretty good, but was a bit limited due to the use of the zoxide tool.
In short, it had problems switching to the correct folder when it had similar names, like the same prefix.
Primeagen's script is better because it takes into account less folders, but you have to manually set the search folders.
Zoxide is still kept in case I want to switch directories without changing my tmux session.